Added RuleResult to be able to rank results.
The eliminates the need to pull base rules out of iterations. It also handles the conflict between the doublet onset rule and the s onset rule. Now both rules are checked and the better result is taken. However, I'm not sure if the 'better' result will be the same for future conflicting rules.